requirements.txt
godaddypy
requests
PyNamecheap
